# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/profile/me": {
      "get": {
        "operationId": "ProfileController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Профиль текущего пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить профиль текущего авторизованного пользователя",
        "tags": ["Profile"]
      },
      "patch": {
        "operationId": "ProfileController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Тело запроса, валидируется соответствующей zod-схемой.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновленный профиль пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации входных данных"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить профиль текущего пользователя",
        "tags": ["Profile"]
      },
      "delete": {
        "operationId": "ProfileController_deleteMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Аккаунт успешно деактивирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Деактивировать аккаунт текущего пользователя",
        "tags": ["Profile"]
      }
    },
    "/api/v1/profile/avatar": {
      "post": {
        "operationId": "ProfileController_uploadAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Аватар успешно загружен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный файл аватара"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Загрузить аватар профиля",
        "tags": ["Profile"]
      },
      "delete": {
        "operationId": "ProfileController_deleteAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Аватар успешно удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarDeleteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить аватар профиля",
        "tags": ["Profile"]
      }
    },
    "/api/v1/profile/restore": {
      "post": {
        "operationId": "ProfileController_restoreMyProfile",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Аккаунт успешно восстановлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Восстановить деактивированный аккаунт",
        "tags": ["Profile"]
      }
    },
...
  },
  "info": {
    "title": "Mock Interview AI API",
    "description": "Auth API: register, login, logout.\n\nRefresh token передаётся только через HttpOnly cookie\n(HttpOnly; SameSite=Lax; Path=/api/v1/auth; Max-Age=JWT_REFRESH_EXPIRATION)\nи не возвращается в JSON response.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "displayName": {
            "nullable": true,
            "type": "string"
          },
          "username": {
            "nullable": true,
            "type": "string"
          },
          "avatarUrl": {
            "nullable": true,
            "type": "string"
          },
          "telegramUsername": {
            "nullable": true,
            "type": "string"
          },
          "gitUrl": {
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "anyOf": [
              {},
              {
                "type": "string"
              }
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "updatedAt": {
            "anyOf": [
              {},
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "id",
          "displayName",
          "username",
          "avatarUrl",
          "telegramUsername",
          "gitUrl",
          "createdAt",
          "email",
          "updatedAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "username": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{3,30}$"
          },
          "avatarUrl": {
            "nullable": true,
            "type": "string",
            "format": "uri"
          },
          "telegramUsername": {
            "nullable": true,
            "type": "string",
            "pattern": "^@?[a-zA-Z0-9_]{5,32}$"
          },
          "gitUrl": {
            "nullable": true,
            "type": "string",
            "pattern": "^https?:\\/\\/(www\\.)?(github|gitlab)\\.com\\/[a-zA-Z0-9_.-]+\\/?$"
          }
        }
      },
      "AvatarUploadResponseDto": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "string"
          }
        },
        "required": ["avatarUrl"]
      },
      "AvatarDeleteResponseDto": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["avatarUrl"]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
      "RestoreProfileResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/UserProfileDto"
          }
        },
        "required": ["message", "profile"]
      },
...
    }
  }
}